+++ egs/musdb18/d3net/test.sh
@@ -43,7 +43,7 @@ fi
 
 model_choice="last"
 
-model_dir="${save_dir}/model/"
+model_dir="${save_dir}/model"
 log_dir="${save_dir}/log/test"
 json_dir="${save_dir}/json"
