+++ egs/musdb18/open-unmix/test.sh
@@ -48,7 +48,7 @@ fi
 
 model_choice="best"
 
-model_dir="${save_dir}/model/"
+model_dir="${save_dir}/model"
 log_dir="${save_dir}/log/test"
 json_dir="${save_dir}/json"
